+++ core/lib/types/src/operations/forced_exit.rs
@@ -74,7 +74,7 @@ impl ForcedExitOp {
                 .ok_or(ForcedExitOpError::CannotGetInitiatorAccountId)?;
         let target_account_id = u32::from_bytes(&bytes[target_account_id_offset..token_id_offset])
             .ok_or(ForcedExitOpError::CannotGetTargetAccountId)?;
-        let token = u16::from_bytes(&bytes[token_id_offset..amount_offset])
+        let token = u32::from_bytes(&bytes[token_id_offset..amount_offset])
             .ok_or(ForcedExitOpError::CannotGetTokenId)?;
         let amount = BigUint::from_u128(
             u128::from_bytes(&bytes[amount_offset..amount_offset + BALANCE_BIT_WIDTH / 8])
